+++ test/td_dd_web/controllers/data_structure_controller_test.exs
@@ -187,11 +187,28 @@ defmodule TdDdWeb.DataStructureControllerTest do
   end
 
   describe "search" do
-    @tag :admin_authenticated
     setup [:create_data_structure]
 
+    @tag :admin_authenticated
     test "search_all", %{conn: conn, data_structure: %DataStructure{id: id}} do
       conn = post(conn, Routes.data_structure_path(conn, :search), %{})
+
+      assert json = json_response(conn, 200)
+      assert [item] = json["data"]
+      assert filters = json["filters"]
+
+      assert Map.get(item, "id") == id
+    end
+
+    @tag :admin_authenticated
+    test "search with query performs ngram search on name", %{conn: conn} do
+      %{data_structure_id: id} =
+        insert(:data_structure_version,
+          name: "foo_bar_baz",
+          data_structure: build(:data_structure, external_id: "foobarbaz")
+        )
+
+      conn = post(conn, Routes.data_structure_path(conn, :search), %{"query" => "obar"})
 
       assert json = json_response(conn, 200)
       assert [item] = json["data"]