+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-2144] Support ngram-search in structure name
+
 ## [3.8.0] 2019-10-14
 
 ### Fixed